added nfrozModes and reset, also all the tinker files

+++ CoordUtil.h
@@ -18,7 +18,7 @@ using namespace std;
 class CoordUtil {
 public:
 	CoordUtil();
-	CoordUtil(double, double, vector< vector< vector<double> > >, vector<double>, vector<double>, vector< vector<double> >, vector<string>, vector<int>, vector< vector<int> >, string, string);
+	CoordUtil(int, int, vector< vector< vector<double> > >, vector<double>, vector<double>, vector< vector<double> >, vector<string>, vector<int>, vector< vector<int> >, string, string);
 	virtual ~CoordUtil();
 
 	vector< vector<double> > normalModeToCart(vector<Particle>);
